requirement.txt
openai-whisper
watchdog
torch
torchaudio